refactor: steps toward a clean controller, creating common functions into services layers and doing some micro optimizations
