+++ src/main/java/com/tuneurl/webrtc/util/controller/AudioStreamController.java
@@ -91,279 +91,6 @@ public class AudioStreamController extends BaseController {
   }
 
   /**
-   * Convert AudioStreamDatabase to AudioStreamDataResponse.
-   *
-   * @param signature String
-   * @param logger MessageLogger
-   * @param response AudioStreamDataResponse
-   * @param asDB AudioStreamDatabase
-   * @return AudioStreamDataResponse
-   */
-  private AudioStreamDataResponse convertAudioStreamDatabase(
-      final String signature,
-      MessageLogger logger,
-      AudioStreamDataResponse response,
-      AudioStreamDatabase asDB) {
-
-    Long duration = asDB.getAsDuration();
-    final String fiveSecondAudioUrl = asDB.getAsFilename() + ".wav";
-    final String finalAudioStreamUrl = asDB.getAsFilename() + duration.toString() + ".wav";
-
-    response.setFiveSecondAudioUrl(super.getStreamAudioUrlPrefix(fiveSecondAudioUrl));
-    response.setFinalAudioStreamUrl(super.getStreamAudioUrlPrefix(finalAudioStreamUrl));
-
-    response.setConversionId(asDB.getAsId());
-    response.setDuration(duration);
-
-    response.setStatus(asDB.getAsStatus());
-    logger.logExit(signature, new Object[] {response});
-    return response;
-  }
-
-  /**
-   * Set default values for AudioStreamDataResponse members.
-   *
-   * @param signature String
-   * @param logger MessageLogger
-   * @param response AudioStreamDataResponse
-   * @param crc32 String
-   * @return AudioStreamDataResponse
-   */
-  private AudioStreamDataResponse resetResponseValue(
-      final String signature,
-      MessageLogger logger,
-      AudioStreamDataResponse response,
-      final String crc32) {
-    final String url = super.getStreamAudioUrlPrefix(crc32 + ".wav");
-    response.setFiveSecondAudioUrl(url);
-    response.setFinalAudioStreamUrl(url);
-
-    response.setConversionId(0L);
-    response.setDuration(0L);
-
-    response.setStatus(Constants.AUDIOSTREAM_STATUS_FINAL);
-    logger.logExit(signature, new Object[] {response});
-    return response;
-  }
-
-  /**
-   * Update status value.
-   *
-   * @param forceUpdate boolean
-   * @param response AudioStreamDataResponse
-   * @param status Integer
-   * @param asDB AudioStreamDatabase
-   */
-  private void updateStatus(
-      boolean forceUpdate,
-      AudioStreamDataResponse response,
-      Integer status,
-      AudioStreamDatabase asDB) {
-    LocalDateTime localDate;
-    MessageLogger logger = super.getMessageLogger();
-    if (forceUpdate) {
-      asDB.setAsStatus(status);
-      localDate = CommonUtil.asLocalDateTime(new Date());
-      asDB.setAsModified(localDate);
-      asDB = audioStreamService.saveAudioStreamDatabase(asDB, logger);
-
-    } else if (asDB != null) {
-      if (!status.equals(asDB.getAsStatus())) {
-        asDB.setAsStatus(status);
-        localDate = CommonUtil.asLocalDateTime(new Date());
-        asDB.setAsModified(localDate);
-        asDB = audioStreamService.saveAudioStreamDatabase(asDB, logger);
-      }
-    }
-    response.setStatus(status);
-    if (asDB != null) {
-      response.setConversionId(asDB.getAsId());
-    }
-  }
-
-  private String updateWaveFileExtension(final String fileName) {
-    if (fileName.endsWith(".wav")) {
-      return fileName;
-    }
-    return fileName + ".wav";
-  }
-
-  /**
-   * Run WebRTC Script.
-   *
-   * @param signature String
-   * @param isExecute boolean
-   * @param command String
-   * @param url String
-   * @param duration Long
-   * @return AudioStreamDataResponse
-   */
-  private AudioStreamDataResponse runWebRtcScript(
-      final String signature,
-      boolean isExecute,
-      final String command,
-      final String url,
-      final Long duration,
-      AudioStreamDatabase pDB)
-      throws BaseServiceException {
-    AudioStreamDatabase asDB = null;
-    AudioStreamDataResponse response = new AudioStreamDataResponse();
-    Map<String, String> results = null;
-    MessageLogger logger = super.getMessageLogger();
-    // 1. Create a unique temporary filename.
-    String uniqueName = ProcessHelper.createUniqueFilename();
-    String outputFilename = String.format("/tmp/%s.txt", uniqueName);
-    // 2.a Make sure we will create a unique file name.
-    if (ProcessHelper.isFileExist(outputFilename)) {
-      uniqueName = ProcessHelper.createUniqueFilename();
-      outputFilename = String.format("/tmp/%s.txt", uniqueName);
-    }
-    String executionMode = isExecute ? "TRUE" : "FALSE";
-    final String crc32Data = url + '-' + duration.toString();
-    String crc32 = ProcessHelper.genCrc32(crc32Data);
-    Integer status;
-    logger.logEntry(
-        signature,
-        new Object[] {
-          "EXEC=", executionMode,
-          "FNAME=", uniqueName,
-          "CMD=", command,
-          "URL=", url,
-          "TIME=", duration,
-          "CRC32=", crc32
-        });
-
-    // 2.b if isExecute is false, don't check the DB but check the status of wget
-    // 2.c If this crc32 already in DB, return immediately.
-    if (pDB != null) {
-      asDB = pDB;
-    } else {
-      try {
-        asDB = audioStreamService.getAudioStreamDatabaseByNameByDuration(crc32, duration);
-      } catch (BaseServiceException ignore) {
-        asDB = null;
-      }
-    }
-
-    // 2.d check if it is time to kill wget.
-    if (null != asDB) {
-      status = Constants.AUDIOSTREAM_STATUS_FINAL;
-      if (status.equals(asDB.getAsStatus())) {
-        return convertAudioStreamDatabase(signature, logger, response, asDB);
-      }
-      if (ProcessHelper.isConversionExpired(asDB.getAsCreated(), duration)) {
-        executionMode = "KILL";
-      }
-    }
-
-    // 3. get path where to save the audio files
-    String rootDir = super.getSaveAudioFilesFolder(null);
-
-    // 4. Execute the script that uses sleep, wget, and ffmpeg to convert the audio stream into a
-    // 10240 .wav file.
-    // 5. If the audio file exist, it will return immediately
-    // 6. If the audio does not exist, the script will download at least 5 seconds of the audio
-    // 7. create the outputFilename and write configuration on it
-    // 8. crc32 value is the value saved into DB table audio_stream_data.as_filename
-    ProcessBuilder processBuilder =
-        new ProcessBuilder(
-            "/bin/bash",
-            command,
-            uniqueName,
-            outputFilename,
-            crc32,
-            rootDir,
-            url,
-            duration.toString(),
-            executionMode);
-    processBuilder.redirectErrorStream(true);
-    processBuilder.directory(new File(rootDir));
-    Process process;
-    try {
-      process = processBuilder.start();
-      process.waitFor();
-    } catch (IOException | InterruptedException ex) {
-      // 9. On error, return connection ID as zero.
-      logger.logExit(signature, "processBuilder.start() Failed: " + ex.getMessage());
-      ex.printStackTrace();
-      ProcessHelper.deleteFile(outputFilename);
-      return resetResponseValue(signature, logger, response, crc32);
-    }
-    // 10. Extract audio stream conversion status after running run_webrtc_script.sh
-    response.setConversionId(0L);
-    response.setDuration(duration);
-    status = Constants.AUDIOSTREAM_STATUS_INIT;
-    results = ProcessHelper.readTextFileAsArray(outputFilename, true);
-    String error = "";
-    for (String key : results.keySet()) {
-      String value = results.get(key);
-      switch (key) {
-        case "FILENAME":
-          // Check if run_webrtc_script.sh setup the correct Filename.
-          if (!crc32.equals(value)) {
-            if (error.length() > 0) error += "\n";
-            error += "Invalid Filename value";
-          }
-          break;
-        case "fiveSecondAudioUrl":
-          response.setFiveSecondAudioUrl(
-              updateWaveFileExtension(super.getStreamAudioUrlPrefix(value)));
-          break;
-        case "finalAudioStreamUrl":
-          response.setFinalAudioStreamUrl(
-              updateWaveFileExtension(super.getStreamAudioUrlPrefix(value)));
-          break;
-        case "duration":
-          response.setDuration(ProcessHelper.parseLong(value, duration));
-          break;
-        case "status":
-          status = ProcessHelper.parseInt(value, Constants.AUDIOSTREAM_STATUS_INIT);
-          break;
-        case "ERROR":
-          if (!"NONE".equals(value)) {
-            if (error.length() > 0) error += "\n";
-            error += value;
-          }
-          break;
-      }
-    }
-
-    if (error.length() > 0) {
-      CommonUtil.BadRequestException(error);
-      /** NOTREACH */
-    }
-    if (executionMode.equals("KILL")) {
-      status = Constants.AUDIOSTREAM_STATUS_FINAL;
-      updateStatus(true, response, status, asDB);
-    } else if (isExecute) {
-      if (asDB == null) {
-        // 11. Save conversion into audio_stream_data table.
-        asDB = audioStreamService.createAudioStreamDatabase();
-        // 12. Set the unique filename based from the CRC32 of the given URL.
-        asDB.setAsFilename(crc32);
-        asDB.setAsUrl(url);
-        // asDB.setAsDuration(response.getDuration());
-        asDB.setAsDuration(duration);
-        // 13. Status becomes Constants.AUDIOSTREAM_STATUS_FINAL if audio stream conversion is
-        // completed.
-        asDB.setAsStatus(status);
-        LocalDateTime localDate = CommonUtil.asLocalDateTime(new Date());
-        asDB.setAsCreated(localDate);
-        asDB.setAsModified(localDate);
-        // 14. Commit the new audio stream conversion
-        updateStatus(true, response, status, asDB);
-      } else {
-        updateStatus(false, response, status, asDB);
-      }
-    } else {
-      updateStatus(false, response, status, asDB);
-    }
-
-    logger.logExit(signature, new Object[] {response.toString()});
-    return response;
-  }
-
-  /**
    * Access the Wave file.
    *
    * <p>If x-audio-return header exist and it's value is <code>octet</code>,
@@ -408,7 +135,7 @@ public class AudioStreamController extends BaseController {
       CommonUtil.NotFoundException("Invalid file '" + pWaveFile + "'");
       /*NOTREACH*/
     }
-    final String wavePath = super.getSaveAudioFilesFolder(pWaveFile);
+    final String wavePath = audioStreamBaseService.getSaveAudioFilesFolder(pWaveFile);
     FileInputStream fileInput = null;
     try {
       fileInput = new FileInputStream(wavePath);
@@ -521,11 +248,11 @@ public class AudioStreamController extends BaseController {
     }
 
     /* 4. Assemble path where the bash shell reside: /home/ubuntu/audio/run_webrtc_script.sh */
-    final String command = super.getSaveAudioFilesFolder(Constants.RUN_WEBRTC_SCRIPT);
+    final String command = audioStreamBaseService.getSaveAudioFilesFolder(Constants.RUN_WEBRTC_SCRIPT);
 
     // 5. Run the script.
     AudioStreamDataResponse response =
-        runWebRtcScript(signature, true, command, url, duration, null);
+        audioStreamBaseService.runWebRtcScript(signature, true, command, url, duration, null);
 
     // 6. Return the result
     return ResponseEntity.ok().body(response);
@@ -618,11 +345,11 @@ public class AudioStreamController extends BaseController {
     AudioStreamDatabase asdb = audioStreamService.getAudioStreamDatabaseById(pConversionid);
 
     /* 4. Assemble path where the bash shell reside: /home/ubuntu/audio/run_webrtc_script.sh */
-    final String command = super.getSaveAudioFilesFolder(Constants.RUN_WEBRTC_SCRIPT);
+    final String command = audioStreamBaseService.getSaveAudioFilesFolder(Constants.RUN_WEBRTC_SCRIPT);
 
     // 5. Run the script.
     AudioStreamDataResponse response =
-        runWebRtcScript(signature, false, command, asdb.getAsUrl(), asdb.getAsDuration(), asdb);
+            audioStreamBaseService.runWebRtcScript(signature, false, command, asdb.getAsUrl(), asdb.getAsDuration(), asdb);
     response.setConversionId(asdb.getAsId());
 
     return ResponseEntity.ok().body(response);
@@ -889,7 +616,7 @@ public class AudioStreamController extends BaseController {
     // The Fingerprint rate.
     Long fingerprintRate = audioDataEntry.getFingerprintRate();
     // Root dir
-    String rootDir = super.getSaveAudioFilesFolder(null);
+    String rootDir = audioStreamBaseService.getSaveAudioFilesFolder(null);
     logger.logEntry(
         signature,
         new Object[] {
@@ -966,199 +693,22 @@ public class AudioStreamController extends BaseController {
       @Valid @RequestBody EvaluateAudioStreamEntry evaluateAudioStreamEntry,
       HttpServletRequest httpRequest,
       HttpServletResponse httpResponse) {
+
+
     final String signature = "evaluateAudioStream";
     // final String signature2 = "evaluateAudioStream:inner";
     super.saveAnalytics(signature, httpRequest);
+
     MessageLogger logger = super.getMessageLogger();
     AudioDataEntry audioDataEntry = evaluateAudioStreamEntry.getAudioData();
-    // The Audio Stream URL.
-    String url = CommonUtil.getString(audioDataEntry.getUrl(), Constants.AUDIOSTREAM_URL_SIZE);
-    // The Data.
-    short[] data = audioDataEntry.getData();
-    // The Size of data.
-    int size = audioDataEntry.getSize().intValue();
-    // The Sample rate.
-    Long sampleRate = audioDataEntry.getSampleRate();
-    // The Duration.
-    Long duration = audioDataEntry.getDuration();
-    // The Fingerprint rate.
-    Long fingerprintRate = audioDataEntry.getFingerprintRate();
-    // The Triggersound Fingerprint Data.
-    byte[] dataFingerprint = evaluateAudioStreamEntry.getDataFingerprint();
-    // The size of Fingerprint Data.
-    Long sizeFingerprint = evaluateAudioStreamEntry.getSizeFingerprint();
-    logger.logEntry(
-        signature,
-        new Object[] {
-          "url=", url,
-          "data=", data.length == size,
-          "size=", size,
-          "SRate=", sampleRate,
-          "duration=", duration,
-          "FRate=", fingerprintRate,
-          "fingerprintData=", dataFingerprint.length == sizeFingerprint,
-          "sizeFingerprint=", sizeFingerprint
-        });
 
     // 1. Check for ADMIN or USER role.
     if (!super.canAccessAudioWithoutLogin()) {
       super.getSdkClientCredentials(signature, UserType.LOGIN_FOR_USER, httpRequest, httpResponse);
     }
-    Converter.checkAudioDataEntryDataSize(audioDataEntry);
-    Converter.validateShortDataSize(data, size);
-    Converter.validateDataSizeEx(dataFingerprint, sizeFingerprint.intValue());
-    Converter.validateDurationEx(duration);
-    final String fileName = Converter.validateUrlOrGencrc32(url);
-    ProcessHelper.checkNullOrEmptyString(fileName, "AudioDataEntry.Url");
-    if (duration < 1L || duration > 480L) {
-      CommonUtil.BadRequestException("Duration must be 1 to 480 seconds only");
-    }
-    EvaluateAudioStreamResponse response = new EvaluateAudioStreamResponse();
-    List<TuneUrlTag> liveTags = new ArrayList<TuneUrlTag>();
-    long elapse;
-    long timeOffset, baseOffset;
-    long iStart, iEnd;
-    long maxDuration = Converter.muldiv(1000, duration - 6L, 1L);
-    long count, counts = Converter.muldiv(1000, duration - 6L, 100);
-    int dSize;
-    short[] dData;
-    FingerprintCompareResponse fcr = null;
-    FingerprintCompareResponse fca;
-    FingerprintCompareResponse fcb;
-    FingerprintCompareResponse fcc;
-    FingerprintCompareResponse fcd;
-    FingerprintCompareResponse fce;
-    TuneUrlTag tag;
-    boolean isDebugOn = Constants.DEBUG_FINGERPRINTING;
-    String rootDir = super.getSaveAudioFilesFolder(null);
-    String debugUniqueName = ProcessHelper.createUniqueFilename();
-    String debugDir = String.format("%s/%s", rootDir, "debug");
-    String rootDebugDir = String.format("%s/%s", debugDir, debugUniqueName);
-    FingerprintResponse fr = null;
-    FingerprintResponse audioFr = null;
-    if (isDebugOn) {
-      ProcessHelper.makeDir(debugDir);
-    }
-    Random random = new Random();
-    random.setSeed(new Date().getTime());
-    for (count = 0L, elapse = 0L; count < counts && elapse < maxDuration; count++, elapse += 100L) {
 
-      FingerprintCollection result =
-          FingerprintUtility.collectFingerprint(
-              logger,
-              rootDir,
-              data,
-              elapse,
-              random,
-              fingerprintRate,
-              dataFingerprint,
-              Constants.FINGERPRINT_INCREMENT_DELTA);
+    EvaluateAudioStreamResponse response = audioStreamBaseService.evaluateAudioStream(audioDataEntry, evaluateAudioStreamEntry, signature);
 
-      List<FingerprintResponse> frSelection = result.getFrCollection();
-      List<FingerprintCompareResponse> selection = result.getFcrCollection();
-
-      timeOffset = elapse;
-      fcr = null;
-      fr = null;
-      if (selection.size() == 5) {
-        fca = selection.get(0);
-        fcb = selection.get(1);
-        fcc = selection.get(2);
-        fcd = selection.get(3);
-        fce = selection.get(4);
-
-        //  8: N P N N N => P is the valid TuneUrl trigger sound
-        // 15: N P P P P => N is the valid TuneUrl trigger sound
-        // 30: P P P P N => N is the valid TuneUrl trigger sound
-        if (FingerprintUtility.hasNegativeFrameStartTimeEx(fca)
-            && FingerprintUtility.hasPositiveFrameStartTimeEx(fcb)) {
-          if (FingerprintUtility.hasNegativeFrameStartTimeEx(fcc)) {
-            // N P N
-            if (FingerprintUtility.isFrameStartTimeEqual(fca, fcc)
-                && FingerprintUtility.isFrameStartTimeEqual(fcc, fcd)
-                && FingerprintUtility.isFrameStartTimeEqual(fcd, fce)) {
-              // N P N N N => P is the valid TuneUrl trigger sound
-              fcr = selection.get(1);
-              fr = frSelection.get(1);
-            }
-          } else if (FingerprintUtility.hasPositiveFrameStartTimeEx(fcc)
-              && FingerprintUtility.isFrameStartTimeEqual(fcc, fcb)
-              && FingerprintUtility.isFrameStartTimeEqual(fcc, fcd)
-              && FingerprintUtility.isFrameStartTimeEqual(fcd, fce)) {
-            // N P P P P => N is the valid TuneUrl trigger sound
-            fcr = selection.get(0);
-            fr = frSelection.get(0);
-          }
-        } else if (FingerprintUtility.hasPositiveFrameStartTimeEx(fca)
-            && FingerprintUtility.hasNegativeFrameStartTimeEx(fce)) {
-          // P . . . N
-          if (FingerprintUtility.isFrameStartTimeEqual(fca, fcb)
-              && FingerprintUtility.isFrameStartTimeEqual(fcb, fcc)
-              && FingerprintUtility.isFrameStartTimeEqual(fcc, fcd)) {
-            // P P P P N => N is the valid TuneUrl trigger sound
-            fcr = selection.get(4);
-            fr = frSelection.get(4);
-          }
-        }
-
-        if (fcr != null) {
-
-          // Grab the audio after the triggersound
-          timeOffset = fcr.getOffset();
-          baseOffset = timeOffset;
-
-          timeOffset = timeOffset + 1000L;
-          iStart = Converter.muldiv(timeOffset, fingerprintRate, 1000L);
-          iEnd = Converter.muldiv(timeOffset + 5000L, fingerprintRate, 1000L);
-          dSize = (int) (iEnd - iStart);
-          dData = Converter.convertListShortEx(data, (int) iStart, dSize);
-          if (dData == null) break;
-          // Calculate the audio's fingerprint
-          // audioFr = ExternalCppModules.calculateFingerprint(null, dData, dData.length);
-          audioFr =
-              FingerprintUtility.runExternalFingerprinting(
-                  random, logger, rootDir, dData, dData.length);
-
-          tag = FingerprintUtility.newTag(false, 0L, audioFr, fcr);
-          liveTags.add(tag);
-
-          if (isDebugOn) {
-            FingerprintUtility.saveAudioClipsAt(
-                logger,
-                true,
-                evaluateAudioStreamEntry,
-                fr,
-                fcr,
-                "trigger",
-                baseOffset,
-                1000L,
-                rootDebugDir,
-                debugUniqueName);
-
-            FingerprintUtility.saveAudioClipsAt(
-                logger,
-                false,
-                evaluateAudioStreamEntry,
-                audioFr,
-                fcr,
-                "audio",
-                baseOffset + 1000L,
-                5000L,
-                rootDebugDir,
-                debugUniqueName);
-          } // if (isDebugOn)
-        } // if (fcr != null)
-      } // if (selection.size() == 5)
-    } // for (...)
-    liveTags = FingerprintUtility.pruneTags(liveTags);
-    counts = (long) liveTags.size();
-    response.setTagCounts(counts);
-    response.setLiveTags(liveTags);
-    response.setTuneUrlCounts((long) liveTags.size());
-    if (isDebugOn) {
-      FingerprintUtility.displayLiveTags(signature, logger, liveTags);
-    }
-    logger.logExit(signature, new Object[] {"counts=", counts, "liveTags.size", liveTags.size()});
     return ResponseEntity.ok().body(response);
   }
 }
